requirements.txt
streamlit
COHERE_API_KEY
pdfplumber
python-dotenv
